WFLY-16038 Extract MSC-related classes within wildfly-clustering-infinispan-client into separate module
