+++ clustering/infinispan/client/service/src/main/java/org/wildfly/clustering/infinispan/client/service/InfinispanClientRequirement.java
@@ -1,0 +1,63 @@
+/*
+ * JBoss, Home of Professional Open Source.
+ * Copyright 2018, Red Hat, Inc., and individual contributors
+ * as indicated by the @author tags. See the copyright.txt file in the
+ * distribution for a full listing of individual contributors.
+ *
+ * This is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of
+ * the License, or (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this software; if not, write to the Free
+ * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+ * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+ */
+
+package org.wildfly.clustering.infinispan.client.service;
+
+import org.infinispan.client.hotrod.RemoteCacheContainer;
+import org.infinispan.client.hotrod.configuration.Configuration;
+import org.jboss.as.clustering.controller.UnaryRequirementServiceNameFactory;
+import org.jboss.as.clustering.controller.UnaryServiceNameFactory;
+import org.jboss.as.clustering.controller.UnaryServiceNameFactoryProvider;
+import org.wildfly.clustering.service.UnaryRequirement;
+
+/**
+ * @author Paul Ferraro
+ */
+public enum InfinispanClientRequirement implements UnaryRequirement, UnaryServiceNameFactoryProvider {
+
+    REMOTE_CONTAINER("org.wildfly.clustering.infinispan.remote-cache-container", RemoteCacheContainer.class),
+    REMOTE_CONTAINER_CONFIGURATION("org.wildfly.clustering.infinispan.remote-cache-container-configuration", Configuration.class),
+    ;
+    private final String name;
+    private final Class<?> type;
+    private final UnaryServiceNameFactory factory = new UnaryRequirementServiceNameFactory(this);
+
+    InfinispanClientRequirement(String name, Class<?> type) {
+        this.name = name;
+        this.type = type;
+    }
+
+    @Override
+    public String getName() {
+        return this.name;
+    }
+
+    @Override
+    public Class<?> getType() {
+        return this.type;
+    }
+
+    @Override
+    public UnaryServiceNameFactory getServiceNameFactory() {
+        return this.factory;
+    }
+}
