+++ clustering/infinispan/client/spi/src/main/java/org/wildfly/clustering/infinispan/client/near/SimpleKeyWeigher.java
