--- clustering/infinispan/client/src/main/java/org/wildfly/clustering/infinispan/client/NearCacheFactory.java
@@ -1,46 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source.
- * Copyright 2020, Red Hat, Inc., and individual contributors
- * as indicated by the @author tags. See the copyright.txt file in the
- * distribution for a full listing of individual contributors.
- *
- * This is free software; you can redistribute it and/or modify it
- * under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation; either version 2.1 of
- * the License, or (at your option) any later version.
- *
- * This software is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this software; if not, write to the Free
- * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
- * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
- */
-
-package org.wildfly.clustering.infinispan.client;
-
-import org.infinispan.client.hotrod.configuration.NearCacheMode;
-import org.infinispan.client.hotrod.event.impl.ClientListenerNotifier;
-import org.infinispan.client.hotrod.near.NearCacheService;
-
-/**
- * Factory for creating near cache services.
- * @author Paul Ferraro
- */
-public interface NearCacheFactory<K, V> {
-    /**
-     * Creates a near cache service using the specified notifier.
-     * @param notifier a notifier for client events
-     * @return a near cache service
-     */
-    NearCacheService<K, V> createService(ClientListenerNotifier notifier);
-
-    /**
-     * Indicates the mode associated with this factory.
-     * @return a near cache mode
-     */
-    NearCacheMode getMode();
-}
