--- ee-feature-pack/common/src/main/resources/modules/system/layers/base/org/wildfly/clustering/infinispan/client/main/module.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-  ~ JBoss, Home of Professional Open Source.
-  ~ Copyright 2010, Red Hat, Inc., and individual contributors
-  ~ as indicated by the @author tags. See the copyright.txt file in the
-  ~ distribution for a full listing of individual contributors.
-  ~
-  ~ This is free software; you can redistribute it and/or modify it
-  ~ under the terms of the GNU Lesser General Public License as
-  ~ published by the Free Software Foundation; either version 2.1 of
-  ~ the License, or (at your option) any later version.
-  ~
-  ~ This software is distributed in the hope that it will be useful,
-  ~ but WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
-  ~ Lesser General Public License for more details.
-  ~
-  ~ You should have received a copy of the GNU Lesser General Public
-  ~ License along with this software; if not, write to the Free
-  ~ Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
-  ~ 02110-1301 USA, or see the FSF site: http://www.fsf.org.
-  -->
-
-<module xmlns="urn:jboss:module:1.9" name="org.wildfly.clustering.infinispan.client">
-
-    <properties>
-        <property name="jboss.api" value="private"/>
-    </properties>
-
-    <resources>
-        <artifact name="${org.wildfly:wildfly-clustering-infinispan-client}"/>
-    </resources>
-
-    <dependencies>
-        <module name="internal.javax.transaction.api"/>
-        <module name="com.github.ben-manes.caffeine"/>
-        <module name="org.infinispan.client.hotrod"/>
-        <module name="org.infinispan.commons"/>
-        <module name="org.jboss.as.clustering.common"/>
-        <module name="org.jboss.as.controller"/>
-        <module name="org.jboss.modules"/>
-        <module name="org.jboss.msc"/>
-        <module name="org.wildfly.clustering.api"/>
-        <module name="org.wildfly.clustering.infinispan.marshalling"/>
-        <module name="org.wildfly.clustering.service"/>
-    </dependencies>
-</module>
