+++ ee-feature-pack/common/src/main/resources/modules/system/layers/base/org/wildfly/clustering/web/hotrod/main/module.xml
@@ -52,7 +52,9 @@
         <module name="org.wildfly.clustering.ee.cache" services="import"/>
         <module name="org.wildfly.clustering.ee.hotrod"/>
         <module name="org.wildfly.clustering.ee.spi"/>
-        <module name="org.wildfly.clustering.infinispan.client"/>
+        <module name="org.wildfly.clustering.infinispan.client.api"/>
+        <module name="org.wildfly.clustering.infinispan.client.service"/>
+        <module name="org.wildfly.clustering.infinispan.client.spi"/>
         <module name="org.wildfly.clustering.marshalling.api"/>
         <module name="org.wildfly.clustering.marshalling.jboss"/>
         <module name="org.wildfly.clustering.marshalling.protostream"/>
